+++ packages/core-js/es/regexp/index.js
@@ -1,5 +1,6 @@
 require('../../modules/es.regexp.constructor');
 require('../../modules/es.regexp.to-string');
+require('../../modules/es.regexp.dot-all');
 require('../../modules/es.regexp.exec');
 require('../../modules/es.regexp.flags');
 require('../../modules/es.regexp.sticky');