+++ packages/core-js/modules/es.regexp.dot-all.js
@@ -1,20 +1,20 @@
 var DESCRIPTORS = require('../internals/descriptors');
-var UNSUPPORTED_Y = require('../internals/regexp-sticky-helpers').UNSUPPORTED_Y;
+var UNSUPPORTED_DOT_ALL = require('../internals/regexp-unsupported-dot-all');
 var defineProperty = require('../internals/object-define-property').f;
 var getInternalState = require('../internals/internal-state').get;
 var RegExpPrototype = RegExp.prototype;
 
-// `RegExp.prototype.sticky` getter
-// https://tc39.es/ecma262/#sec-get-regexp.prototype.sticky
-if (DESCRIPTORS && UNSUPPORTED_Y) {
-  defineProperty(RegExpPrototype, 'sticky', {
+// `RegExp.prototype.dotAll` getter
+// https://tc39.es/ecma262/#sec-get-regexp.prototype.dotall
+if (DESCRIPTORS && UNSUPPORTED_DOT_ALL) {
+  defineProperty(RegExpPrototype, 'dotAll', {
     configurable: true,
     get: function () {
       if (this === RegExpPrototype) return undefined;
       // We can't use InternalStateModule.getterFor because
       // we don't add metadata for regexps created by a literal.
       if (this instanceof RegExp) {
-        return !!getInternalState(this).sticky;
+        return !!getInternalState(this).dotAll;
       }
       throw TypeError('Incompatible receiver, RegExp required');
     }