+++ tests/compat/tests.js
@@ -145,6 +145,12 @@ var TYPED_ARRAY_CONSTRUCTORS_NOT_REQUIRES_WRAPPERS = function () {
     && new Int8Array(new ArrayBuffer(2), 1, undefined).length == 1;
 };
 
+function NCG_SUPPORT() {
+  var re = RegExp('(?<a>b)');
+  return re.exec('b').groups.a === 'b' &&
+    'b'.replace(re, '$<a>c') === 'bc';
+}
+
 function createIsRegExpLogicTest(name) {
   return function () {
     var regexp = /./;
@@ -782,7 +788,7 @@ GLOBAL.tests = {
   'es.reflect.to-string-tag': function () {
     return Reflect[Symbol.toStringTag];
   },
-  'es.regexp.constructor': function () {
+  'es.regexp.constructor': [NCG_SUPPORT, function () {
     var re1 = /a/g;
     var re2 = /a/g;
     re2[Symbol.match] = false;
@@ -791,9 +797,13 @@ GLOBAL.tests = {
       && RegExp(re2) !== re2
       && RegExp(re1, 'i') == '/a/i'
       && new RegExp('a', 'y') // just check that it doesn't throw
+      && RegExp('.', 's').exec('\n')
       && RegExp[Symbol.species];
+  }],
+  'es.regexp.dot-all': function () {
+    return RegExp('.', 's').dotAll;
   },
-  'es.regexp.exec': function () {
+  'es.regexp.exec': [NCG_SUPPORT, function () {
     var re1 = /a/;
     var re2 = /b*/g;
     var reSticky = new RegExp('a', 'y');
@@ -806,10 +816,11 @@ GLOBAL.tests = {
       && reSticky.exec('abc')[0] === 'a'
       && reSticky.exec('abc') === null
       && (reSticky.lastIndex = 1, reSticky.exec('bac')[0] === 'a')
-      && (reStickyAnchored.lastIndex = 2, reStickyAnchored.exec('cba') === null);
-  },
+      && (reStickyAnchored.lastIndex = 2, reStickyAnchored.exec('cba') === null)
+      && RegExp('.', 's').exec('\n');
+  }],
   'es.regexp.flags': function () {
-    return /./g.flags === 'g' && new RegExp('a', 'y').flags === 'y';
+    return Object.getOwnPropertyDescriptor(RegExp.prototype, 'flags').get.call({ dotAll: true, sticky: true }) === 'sy';
   },
   'es.regexp.sticky': function () {
     return new RegExp('a', 'y').sticky === true;