+++ tests/tests/es.regexp.constructor.js
@@ -1,5 +1,5 @@
 /* eslint-disable regexp/order-in-character-class, regexp/no-dupe-characters-character-class -- required for testing */
-/* eslint-disable regexp/no-useless-flag, regexp/sort-flags -- required for testing */
+/* eslint-disable regexp/no-useless-character-class, regexp/no-useless-flag, regexp/sort-flags -- required for testing */
 import { DESCRIPTORS, GLOBAL } from '../helpers/constants';
 import { nativeSubclass } from '../helpers/helpers';
 
@@ -42,5 +42,48 @@ if (DESCRIPTORS) {
       assert.ok(new Subclass() instanceof RegExp, 'correct subclassing with native classes #2');
       assert.ok(new Subclass('^abc$').test('abc'), 'correct subclassing with native classes #3');
     }
+  });
+
+  QUnit.test('RegExp dotAll', assert => {
+    assert.same(RegExp('.', '').test('\n'), false, 'dotAll missed');
+    assert.same(RegExp('.', 's').test('\n'), true, 'dotAll basic');
+    assert.same(RegExp('[.]', 's').test('\n'), false, 'dotAll brackets #1');
+    assert.same(RegExp('[.].', '').test('.\n'), false, 'dotAll brackets #2');
+    assert.same(RegExp('[.].', 's').test('.\n'), true, 'dotAll brackets #3');
+    assert.same(RegExp('[[].', 's').test('[\n'), true, 'dotAll brackets #4');
+    assert.same(RegExp('.[.[].\\..', 's').source, '.[.[].\\..', 'dotAll correct source');
+
+    const string = '123\n456789\n012';
+    const re = RegExp('(\\d{3}).\\d{3}', 'sy');
+
+    let match = re.exec(string);
+    assert.same(match[1], '123', 's with y #1');
+    assert.same(re.lastIndex, 7, 's with y #2');
+
+    match = re.exec(string);
+    assert.same(match[1], '789', 's with y #3');
+    assert.same(re.lastIndex, 14, 's with y #4');
+  });
+
+  QUnit.test('RegExp NCG', assert => {
+    assert.same(RegExp('(?<a>b)').exec('b').groups?.a, 'b', 'NCG #1');
+    // eslint-disable-next-line regexp/no-unused-capturing-group -- required for testing
+    assert.same(RegExp('(b)').exec('b').groups, undefined, 'NCG #2');
+    const { groups } = RegExp('foo:(?<foo>\\w+),bar:(?<bar>\\w+)').exec('foo:abc,bar:def');
+    assert.deepEqual(groups, { foo: 'abc', bar: 'def' }, 'NCG #3');
+    // fails in Safari
+    // assert.same(Object.getPrototypeOf(groups), null, 'NCG #4');
+    assert.same('foo:abc,bar:def'.replace(RegExp('foo:(?<foo>\\w+),bar:(?<bar>\\w+)'), '$<bar>,$<foo>'), 'def,abc', 'replace #1');
+    assert.same('foo:abc,bar:def'.replace(RegExp('foo:(?<foo>\\w+),bar:(?<bar>\\w+)'), (...args) => {
+      const { foo, bar } = args.pop();
+      return `${ bar },${ foo }`;
+    }), 'def,abc', 'replace #2');
+    assert.same('12345'.replaceAll(RegExp('(?<d>[2-4])', 'g'), '$<d>$<d>'), '12233445', 'replaceAll');
+    // eslint-disable-next-line no-invalid-regexp -- required for testing
+    assert.throws(() => RegExp('(?<1a>b)'), SyntaxError, 'incorrect group name #1');
+    // eslint-disable-next-line no-invalid-regexp -- required for testing
+    assert.throws(() => RegExp('(?<a#>b)'), SyntaxError, 'incorrect group name #2');
+    // eslint-disable-next-line no-invalid-regexp -- required for testing
+    assert.throws(() => RegExp('(?< a >b)'), SyntaxError, 'incorrect group name #3');
   });
 }